Factor RoomFacePile into a separate file
